refactor(web): split PlayerProfile into folder, adopt useStaleRefresh
