+++ apps/web/src/components/player/PlayerProfile/StaleBadge.tsx
@@ -1,0 +1,12 @@
+'use client'
+
+import { Badge } from '@brawltome/ui'
+
+export function StaleBadge() {
+  return (
+    <Badge variant="secondary" className="gap-2 animate-pulse">
+      <div className="w-2 h-2 bg-primary rounded-full animate-ping" />
+      Syncing live data...
+    </Badge>
+  )
+}
